--- .eslintrc.json
@@ -1,26 +1,0 @@
-{
-  "extends": "brightspace/lit-config",
-  "globals": {
-    "Prism": false
-  },
-  "overrides": [{
-    "files": [
-      "**/demo/**.html"
-    ],
-    "rules": {
-      "no-console": 0
-    }
-  },
-  {
-    "files": "mixins/demo/**.html",
-    "rules": {
-      "quotes": 0,
-      "no-undef": 0,
-      "indent": 0
-    }
-  },
-  {
-    "files": "**/test/*",
-    "extends": "brightspace/testing-config"
-  }]
-}
