+++ package.json
@@ -47,9 +47,9 @@
   "devDependencies": {
     "@brightspace-ui/stylelint-config": "^1",
     "@brightspace-ui/testing": "^1",
-    "@eslint/compat": "^1.2.1",
-    "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.13.0",
+    "@eslint/compat": "^1",
+    "@eslint/eslintrc": "^3",
+    "@eslint/js": "^9",
     "@rollup/plugin-dynamic-import-vars": "^2",
     "@rollup/plugin-node-resolve": "^15",
     "@rollup/plugin-replace": "^6",
